New Server Class
I started this framework because the other server class could not pass data from one client to the other. That is because the other server class handles clients individually and not together. This new server class handles SESSIONS of 2 players and will allow us to create multiple input/output streams in the same session. This will also allow us to disable buttons and such for the turn based portion. As of right now, the code is not running but there are the correct comments where the old code needs to be inserted. I need to leave now but I will definitely have this server class done and running by the end of the night. - Kevin Hale 18:28 11-25-2016